XMLHttpRequest/tests/submissions/Opera/abort-event-listeners.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: abort() should not reset event listeners</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest(),
            listener = function() {}
        client.onreadystatechange = listener
        client.open("GET", "resources/delay.php?ms=200")
        client.send(null)
        client.abort()
        assert_equals(client.onreadystatechange, listener)
      })
      test.done()
    </script>
  </body>
</html>
